requirements.txt
google-generativeai>=0.3.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pillow>=8.3.0
python-dotenv>=0.19.0
opencv-python>=4.5.0